Apply routes->services->repositories layering to templates (#640, part 2)
